Small changes to client
