+++ src/main/resources/central_server.properties
@@ -1,0 +1,10 @@
+spring.config.name=central_server
+server.url=http://localhost:8090
+server.port=8090
+server.api.url=${server.url}/api
+server.api.register.url=${server.api.url}/register
+server.api.login.url=${server.api.url}/login
+server.api.logout.url=${server.api.url}/logout
+server.api.users.url=${server.api.url}/users
+server.api.online_users.url=${server.api.url}/online_users
+spring.application.admin.jmx-name=org.springframework.boot:type=Admin_Central,name=SpringApplication_Central
