+++ src/main/resources/client_server.properties
@@ -1,0 +1,5 @@
+spring.config.name=client_server
+server.url=http://localhost:8020
+front.port=8020
+central_server_properties=central_server.properties
+spring.application.admin.jmx-name=org.springframework.boot:type=Admin_Client,name=SpringApplication_Client
