--- src/main/resources/main_server.properties
@@ -1,8 +1,0 @@
-spring.config.name=main_server
-server.url=http://localhost:8090
-server.port=8090
-spring.datasource.url=jdbc:postgresql://localhost:9000/central_server_db
-spring.datasource.username=username
-spring.datasource.password=password
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
